% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_expe_and_counter_func.R
\name{add_expe_and_counter_func}
\alias{add_expe_and_counter_func}
\title{add_expe_and_counter_func}
\usage{
add_expe_and_counter_func(my_dataset)
}
\arguments{
\item{my_dataset}{}
}
\value{

}
\description{
Add couter for each experiment.
}
