% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_layer_point_ggplot_func.R
\name{animal_layer_point_ggplot_func}
\alias{animal_layer_point_ggplot_func}
\title{animal_layer_point_ggplot_func}
\usage{
animal_layer_point_ggplot_func(
  ggplot_obj,
  dataset,
  xaxe = "Animal",
  colored_by,
  yaxe,
  jitter_width = 3.5,
  animal_size = 2,
  animal_alpha = 0.4
)
}
\arguments{
\item{ggplot_obj}{A ggplot object. the base canvas plot which you want to add the layer to.}

\item{dataset}{A dataframe object. The dataset for analysis, eg wave kinetics or wave occurrence df.}

\item{xaxe}{A string. The variable you want to compare with. Default to \code{Animal}.}

\item{colored_by}{A string. The criteria to assigning color to the dots. Options possibles are by \code{animal_no} and by \code{animal_type}. Default to \code{animal_type}.}

\item{yaxe}{A string. The variable name of interest to plot.}

\item{jitter_width}{A double. Value assigned to the jitter width. Default to \code{3.5}.}

\item{animal_size}{A double. Value assigned to the size of the dots. Default to \code{2}.}

\item{animal_alpha}{A double. Value assigned to the transparency of the dots. Default to \code{0.4}.}
}
\value{
A ggplot object with a dotplot additional layer of individual Animal datapoints.
}
\description{
Create a dot plot ggplot object layer on a base canvas plot. This function makes a dot plot by calculating the average value for each animal based on the cellular level, so it shows the average wave for each animal. Note: this function is wrapped inside the final superplot function.
}
