% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_layer_point_ggplot_func.R
\name{cell_layer_point_ggplot_func}
\alias{cell_layer_point_ggplot_func}
\title{cell_layer_point_ggplot_func}
\usage{
cell_layer_point_ggplot_func(
  ggplot_obj,
  dataset,
  yaxe,
  my_grouping_vars = c("Animal_No", "Animal", "Condition", "Experiment", "Treatment"),
  jitter_width = 3.5,
  cell_size = 2,
  cell_alpha = 0.4
)
}
\arguments{
\item{ggplot_obj}{A ggplot object. the base canvas plot which you want to add the layer to.}

\item{dataset}{A dataframe object. The dataset for analysis, eg wave kinetics or wave occurrence df.}

\item{yaxe}{A string. The variable name of interest to plot.}

\item{my_grouping_vars}{Character vector. A character vector of groups names assigned to perform the cell aggregation. Don't change at least you know what you are doing!}

\item{jitter_width}{A double. Value assigned to the jitter width. Default to \code{3.5}.}

\item{cell_size}{A double. Value assigned to the size of the dots. Default to \code{2}.}

\item{cell_alpha}{A double. Value assigned to the transparency of the dots. Default to \code{0.4}.}
}
\value{
A ggplot object with a dotplot additional layer of individual cell datapoint.
}
\description{
Create a dot plot ggplot object layer on a base canvas plot. This function make dots plot with aggregated waves (average) by cells, that is, every datapoint represent the average value of multiples waves of single cells. Note: this function is wrapped inside the final superplot function.
}
\examples{
# the example is missing
}
