% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_nested_df.R
\name{describe_nested_df}
\alias{describe_nested_df}
\title{Make summary statistics on nested dataset}
\usage{
describe_nested_df(
  nested_df,
  targed_dataset = "data",
  faceted_by_1 = "Condition",
  faceted_by_2 = "."
)
}
\arguments{
\item{nested_df}{A nested dataset. Ideally, previously grouped by the paramters of interes to be analyzed. Eg. wave frequency, wave latency, amplitude, etc.}

\item{targed_dataset}{Character. the column name of the dataset to perform the summary. By default uses the colum named \code{"data"} containing the raw dataset, but if you have transformed data e.ge logarithmic, centered, etc., you can input the name of such column.}

\item{faceted_by_1}{A string. Grouping variables for visualization. Default to \code{Condition}.}

\item{faceted_by_2}{A string. Aditional grouping variables for visualization. Default to \code{.}.}
}
\value{
A nested dataframe with the additional columns: "Described", "outliers" and "qqplots" for the given input dataframe.
}
\description{
This function compute descriptive statistics on a nested dataframe and create 3 new columns containing: 1) descriptive statistics using the function \code{my_summ_stat_func()}, 2) outliers and extreme values and 3) qqplot to visulaize normality.
}
\examples{
# the example is missing
}
