% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_model_info_func.R
\name{extract_model_info}
\alias{extract_model_info}
\title{extract_model_info}
\usage{
extract_model_info(ns_df)
}
\arguments{
\item{ns_df}{A nested dataframe containing a column with a model object named \code{model}.}
}
\value{
A nested dataframe with additional columns containing model information.
}
\description{
Get information from a model object. This function is a wraper from the {broom} package to glance, tidy and augment the model object. Extract and display also different disgnostic tools for model: eg. LogLik, AIC, BIC.
}
\examples{
# example is missing.
}
