% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_SR_experiment_func.R
\name{filter_SR_experiment_func}
\alias{filter_SR_experiment_func}
\title{Filter permeabilized cells data}
\usage{
filter_SR_experiment_func(
  my_dataset,
  my_date,
  my_Experiment,
  my_Linescan,
  my_Wave_no,
  set_True = NULL
)
}
\arguments{
\item{my_dataset}{A dataframe with raw data of waves.}

\item{my_date}{String. The date of the experiment you want to add or remove.}

\item{my_Experiment}{String. The Experiment with the cell ID indiacting permeabilization and coverslip. eg. \verb{P1 CS1}.}

\item{my_Linescan}{String. The Linescan id. eg. \code{LS1}.}

\item{my_Wave_no}{String. The wave no id. eg. \code{W1}.}

\item{set_True}{Logical. You want to add or remove this experiment?. All experiments in the raw dataframe are by default set to \code{TRUE}.}
}
\value{
A datafranme with the column \code{For_Analysis} modified with the filtered experiments.
}
\description{
This function set to true or False experiment for analysis accordingly. If requires at least a dataset and a date. If other arguments are missing it will set by default all values (Experiments, Linescan, waves) to that one asiggned by the user in the given date.
}
\examples{
# Examples are missing
}
