% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_dirs_depth_n_func.R
\name{list_dirs_depth_n_func}
\alias{list_dirs_depth_n_func}
\title{list_dirs_depth_n_func}
\usage{
list_dirs_depth_n_func(p, n)
}
\arguments{
\item{p}{string. target directory.}

\item{n}{integer. Number of levels to list directories.}
}
\value{
A vector string containing all directories found in the seraching path.
}
\description{
list directories at defined deep steps
}
\examples{
# No exmaples proveided jet
}
