% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sr_data.R
\name{load_sr_data}
\alias{load_sr_data}
\title{Title}
\usage{
load_sr_data(files_path)
}
\arguments{
\item{files_path}{String. the original path where the data is stored. this is only for SR experiments orpermeabilized cells?}
}
\value{
A datatframe with concatenated files of single experiments indicated in the searching path.
}
\description{
Title
}
\examples{
# Example is missing
}
