% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sr_data_func.R
\name{load_sr_data_func}
\alias{load_sr_data_func}
\title{load_sr_data_func}
\usage{
load_sr_data_func(my_dir)
}
\arguments{
\item{my_dir}{String. the original path where the data is stored. this is only for SR experiments orpermeabilized cells?}
}
\value{
A datatframe with concatenated files of single experiments indicated in the searching path.
}
\description{
load_sr_data_func
}
\examples{
# example is missing
}
