% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_multi_profile_func.R
\name{make_multi_profile_func}
\alias{make_multi_profile_func}
\title{make_multi_profile_func}
\usage{
make_multi_profile_func(my_root_img_dir)
}
\arguments{
\item{my_root_img_dir}{string. The root directory containing the linescan image(s). files must end with the following suffix "Smth" and the extention ".tif".}
}
\value{
A dataframe. a rectangular data structure containing the varibales: Time (ms), Fluorescence (in AU), Animal (fct) and Condition (fct)
}
\description{
Make plot profile of multiples  linescan images. Note that the folder containing the images must follow this hierarchical structure: Animal -> Condition.
}
\examples{
# example is missing.
}
