% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_profile_func.R
\name{make_profile_func}
\alias{make_profile_func}
\title{make_profile_func}
\usage{
make_profile_func(my_dir)
}
\arguments{
\item{my_dir}{"A directory with a linescan image in tiff format.}
}
\value{
two vector: "Time" in ms and "Fluorescence" the signal in AU.
}
\description{
Create plot-profile from linescan image
}
\examples{
# example missing.

}
