% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_plot_list_func.R
\name{multi_plot_list_func}
\alias{multi_plot_list_func}
\title{Multiples Plots}
\usage{
multi_plot_list_func(my_dataset, yaxe_vars, my_plot_fun, ...)
}
\arguments{
\item{my_dataset}{A dataframe.}

\item{yaxe_vars}{Character vector of variables names to be plotted.}

\item{my_plot_fun}{Function. A function to be used for generating the plots}

\item{...}{Other parameter to pass.}
}
\value{
A list of multiples ggplot objects.
}
\description{
Plots multiples variables (eg. scatter plots) from a dataframe and a plotting function.
}
\examples{
# The example is still missing...


}
