% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_boxplot_and_jitter_func.R
\name{my_boxplot_and_jitter_func}
\alias{my_boxplot_and_jitter_func}
\title{Boxplot with jitter scatter points}
\usage{
my_boxplot_and_jitter_func(
  dataset,
  xaxe = "Animal",
  yaxe,
  box_color = "Animal",
  scatt_color = box_color,
  faceted_by_1 = "Condition",
  faceted_by_2 = ".",
  jitter_width = 0.5,
  .alpha = 0.25,
  .dot_size = 2.5,
  base_font_size = 22,
  y_limits,
  compare_means = FALSE
)
}
\arguments{
\item{dataset}{A dataframe with raw data of waves.}

\item{xaxe}{String. name of the variable in the x axis.}

\item{yaxe}{String. name of the variable in the y axis.}

\item{box_color}{String. Defining the color for whisker box.}

\item{scatt_color}{String. Defining the color for scatter points.}

\item{faceted_by_1}{String. Define facet for plotting.}

\item{faceted_by_2}{String. Define facet for plotting.}

\item{jitter_width}{Double. Define how big is the with of the scatter points.}

\item{.alpha}{Double. Define transparency for your scatter points.}

\item{.dot_size}{Double. Define the size of the scatter points.}

\item{base_font_size}{A integer. Modify the size of the fonts. Default to 22.}

\item{y_limits}{A double. Optional parameter to set the upper y limit to your plot.}

\item{compare_means}{Logic. Shall the mean of the groups be compared? This perform non-paired test using the Wilcoxon-test. Default to FALSE.}
}
\value{
A ggplot object with boxplot + jitter scatter plot.
}
\description{
This function makes a boxplot and jitter plot for animal comparisons. It needs at least two variable x, y to be plotted.
}
\examples{
# The example is still missing...
}
