% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_cum_occu_wave_plot_func.R
\name{my_cum_occu_wave_plot_func}
\alias{my_cum_occu_wave_plot_func}
\title{my_cum_occu_wave_plot_func}
\usage{
my_cum_occu_wave_plot_func(
  my_dataset,
  my_var,
  reffer_wave_thres,
  line_size = 1,
  reference_line = TRUE,
  ...
)
}
\arguments{
\item{my_dataset}{A dataframe with raw data of waves.}

\item{my_var}{A  character string of length 1. The name of the variable you want to plot the cumulative distribution.}

\item{reffer_wave_thres}{A double. Reference value in seconds indicating the threshold for the calculus of cumulative waves occurrence.}

\item{line_size}{Integer. the size of the lines to display. Default to 1.}

\item{reference_line}{Logic. Should you add a reference vertical line?}

\item{...}{Additional parameters to pass to the aesthetic setting function \code{pptx_presentation_theme_func()}.}
}
\value{
A ggplot object (list) ploting cumulative occurrence of waves in the different animals groups.
}
\description{
plot cum occurence of waves
}
\examples{
# The example is still missing...
}
