% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_occu_bar_plot_func.R
\name{my_occu_bar_plot_func}
\alias{my_occu_bar_plot_func}
\title{my_occu_bar_plot_func}
\usage{
my_occu_bar_plot_func(my_dataset, reffer_wave_thres)
}
\arguments{
\item{my_dataset}{A dataframe with raw data of waves.}

\item{reffer_wave_thres}{A double. Reference value in seconds indicating the threshold for the calculus of cumulative waves occurrence.}
}
\value{
A ggplot object with barplot of waves occurrence in the different animals groups.
}
\description{
bar plot of waves occurrence
}
\examples{
# The example is still missing...
}
