% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_sr_data_func.R
\name{normalize_sr_data_func}
\alias{normalize_sr_data_func}
\title{Normalize to control your values}
\usage{
normalize_sr_data_func(aggre_by_cell_dataset)
}
\arguments{
\item{aggre_by_cell_dataset}{A Dataframe. A previously aggregated by cell dataset.}
}
\value{
A dataframe with nomrlaize numeric variables to control.
}
\description{
This function take a previously aggregated by cell dataset using the function \code{summarize_sr_by_cell_func()} end normalize all numeric variables to the control conditions. The new dataset contain a new variable called \code{Var_type} witch contain the \code{"Normalized"} value or the \code{"Raw"} value.
}
\examples{
# the example is missing
}
