% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summ_stat_func.R, R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\alias{my_summ_stat_func}
\title{Compute summary statistics of multiples variables}
\usage{
my_summ_stat_func(my_dataset)

lhs \%>\% rhs
}
\arguments{
\item{my_dataset}{A dataframe prevously cleaned with the fucntion "lalala".}
}
\value{
compute the mean, median, SD, SEM, number of waves, number of cells, number of animals, and asses Normality of the distribution with the Shapiro-Wilk test.
}
\description{
Compute the summary statistics of the multiples variables from the wave analysis from a previosly cleaned dataset

See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
\examples{
# The example is still missing...
}
\keyword{internal}
