% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_presentation_theme_func.R
\name{pptx_presentation_theme_func}
\alias{pptx_presentation_theme_func}
\title{Plot theme for publication}
\usage{
pptx_presentation_theme_func(base_size = 15)
}
\arguments{
\item{base_size}{An integer. Specify the font and other element's size.}
}
\value{
A plot with modified aesthetics theme.
}
\description{
Modify default ggplot2 object aesthetics for the paper.
}
\details{
This take a ggplot2 object, usually a scatter plot and change the theme to almost totally clear.
}
\examples{
# The example is still missing...

}
