% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_sr_by_cell_func.R
\name{summarize_sr_by_cell_func}
\alias{summarize_sr_by_cell_func}
\title{Sumarize permeabilized cells wave data by cell}
\usage{
summarize_sr_by_cell_func(
  dataset,
  my_grouping_vars = c("Date", "Animal_No", "Animal", "Experiment", "Condition",
    "Treatment")
)
}
\arguments{
\item{dataset}{A dataframe with raw data of waves.}

\item{my_grouping_vars}{Character vector. A character vector of groups names assigned to perform the cell aggregation. Don't change at least you know what you are doing!}
}
\value{
A dataframe with cellular level data agregated, that is, a single point per cell.
}
\description{
This function aggregate (average) multiples waves to single cells level of the permeabilized cells dataset. It computes the cell aggregation to all the variables (numeric columns) in the dataset.
}
\examples{
# the example is missing
}
