% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superplot_func.R
\name{superplot_func}
\alias{superplot_func}
\title{Superplot with violin and scatter plot}
\usage{
superplot_func(
  dataset,
  xaxe,
  yaxe,
  faceted_by_1 = "Condition",
  faceted_by_2 = ".",
  jitter_width = 3.5,
  cell_size = 2,
  cell_alpha = 0.4,
  animal_size = 4,
  animal_alpha = 0.8,
  ...
)
}
\arguments{
\item{dataset}{A dataframe with raw data of waves.}

\item{xaxe}{String. name of the variable in the x axis.}

\item{yaxe}{String. name of the variable in the y axis.}

\item{faceted_by_1}{String. Define facet for plotting.}

\item{faceted_by_2}{String. Define facet for plotting.}

\item{jitter_width}{Double. Define how big is the with of the scatter points.}

\item{cell_size}{Double. define the size of the cell level scatter plot.}

\item{cell_alpha}{Double. define the transparency of the cell level scatter plot.}

\item{animal_size}{Double. define the size of the animal level scatter plot.}

\item{animal_alpha}{Double. define the transparency of the animal level scatter plot.}

\item{...}{Other parameters to pass to the function.}
}
\value{
A ggplot object with boxplot + jitter scatter plot.
}
\description{
Make a violin and scatter plot with aggregated data at different levels (animal, cell, etc).
}
\examples{
# The example is still missing...
}
