% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_layer_point_ggplot_func.R
\name{wave_layer_point_ggplot_func}
\alias{wave_layer_point_ggplot_func}
\title{wave_layer_point_ggplot_func}
\usage{
wave_layer_point_ggplot_func(
  base_ggplot_obj,
  dataset,
  xaxe = "Animal",
  yaxe,
  jitter_width = 3.5,
  wave_size = 2,
  wave_alpha = 0.4
)
}
\arguments{
\item{base_ggplot_obj}{A ggplot object. the base canvas plot which you want to add the layer to.}

\item{dataset}{A dataframe object. The dataset for analysis, eg wave kinetics or wave occurrence df.}

\item{xaxe}{A string. The variable you want to compare with. Default to \code{Animal}.}

\item{yaxe}{A string. The variable name of interest to plot.}

\item{jitter_width}{A double. Value assigned to the jitter width. Default to \code{3.5}.}

\item{wave_size}{A double. Value assigned to the size of the dots. Default to \code{2}.}

\item{wave_alpha}{A double. Value assigned to the transparency of the dots. Default to \code{0.4}.}
}
\value{
A ggplot object with an dotplot additional layer.
}
\description{
Create a dot plot ggplot object layer on a base canvas plot. This function make single waves dot plot with the raw data that is, no aggregation on single waves for multiples cells. Note: this function in wrapped inside the final superplot function.
}
