% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_layer_vio_ggplot_func.R
\name{wave_layer_vio_ggplot_func}
\alias{wave_layer_vio_ggplot_func}
\title{wave_layer_vio_ggplot_func}
\usage{
wave_layer_vio_ggplot_func(
  dataset,
  yaxe,
  xaxe = "Animal",
  faceted_by_1 = "Condition",
  faceted_by_2 = ".",
  y_limits,
  line_size = 1,
  trim = FALSE,
  scaled_to = "width",
  adjusted = 1
)
}
\arguments{
\item{dataset}{A dataframe object. The dataset for analysis, eg wave kinetics or wave occurrence df.}

\item{yaxe}{A string. The variable name of interest to plot.}

\item{xaxe}{A string. The variable you want to compare with. Default to \code{Animal}.}

\item{faceted_by_1}{A string. Grouping variables for visualization. Default to \code{Condition}.}

\item{faceted_by_2}{A string. Aditional grouping variables for visualization. Default to \code{.}.}

\item{y_limits}{A double. Optional paramter to set the upper y limit to your plot.}

\item{line_size}{Integer. the size of the lines to display. Default to 1.}

\item{trim}{Logic. shall the edges of the violin plots be cutted off?. Default to FALSE}

\item{scaled_to}{Character string. If "area" (default), all violins have the same area (before trimming the tails). If "count", areas are scaled proportionally to the number of observations. If "width", all violins have the same maximum width.}

\item{adjusted}{Double. A multiplicate bandwidth adjustment. This makes it possible to adjust the bandwidth while still using the a bandwidth estimator. For example, adjust = 1/2 means use half of the default bandwidth.}
}
\value{
A ggplot object with a violin plot canvas to add additional layers.
}
\description{
Create a base canvas violin ggplot  object. This function make single waves violin plot with the raw data that is, no aggregation on single waves for multiples cells. Note: this function is wrapped inside the final superplot function.
}
\examples{
# the example is missing
}
