% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_expe_and_counter_func.R
\name{add_expe_and_counter_func}
\alias{add_expe_and_counter_func}
\title{add_expe_and_counter_func}
\usage{
add_expe_and_counter_func(my_dataset)
}
\arguments{
\item{my_dataset}{A dataframe containing variables filename.}
}
\value{
the dataframe and a new column named "n_paired" with values 1 for non-paired experiment or 2 pro paired experiments.
}
\description{
Add counter for each experiment to asses paired experiment.
}
\examples{
# The example is still missing...
}
