% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_Treatment_factor_func.R
\name{add_Treatment_factor_func}
\alias{add_Treatment_factor_func}
\title{add_Treatment_factor_func}
\usage{
add_Treatment_factor_func(my_freq_summarized_data)
}
\arguments{
\item{my_freq_summarized_data}{Dataset without not \code{Treatment} factor defined.}
}
\value{
the same dataset with the factor \code{Treatment}.
}
\description{
add_Treatment_factor_func
}
\examples{
# no Example jet provide
}
