% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_and_tidy_df_func.R
\name{clean_and_tidy_df_func}
\alias{clean_and_tidy_df_func}
\title{clean_and_tidy_df_func}
\usage{
clean_and_tidy_df_func(my_df, vars_for_analysis)
}
\arguments{
\item{my_df}{A dataframe with waves raw data. Must contain "Date" and "Experiment" columns.}

\item{vars_for_analysis}{A character vector with variable names to be analyzed.}
}
\value{
A clean dataframe ready for statistics analysis or plotting and add column "n_paired" for paired experiments.
}
\description{
Clean and organize your dataset
}
\examples{
# The example is still missing...
}
