% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_boxplot_and_jitter_func.R
\name{my_boxplot_and_jitter_func}
\alias{my_boxplot_and_jitter_func}
\title{Boxplot with jitter scatter points}
\usage{
my_boxplot_and_jitter_func(
  dataset,
  xaxe = "Animal",
  yaxe,
  box_color = "Animal",
  scatt_color = "Animal",
  faceted_by_1 = "Treatment + Condition",
  faceted_by_2 = ".",
  jitter_width = 0.5,
  .alpha = 0.25,
  .dot_size = 2.5,
  ...
)
}
\arguments{
\item{dataset}{A dataframe.}

\item{xaxe}{String. name of the variable in the x axis.}

\item{yaxe}{String. name of the variable in the y axis.}

\item{box_color}{String. Defining the color for whisker box.}

\item{scatt_color}{String. Defining the color for scatter points.}

\item{faceted_by_1}{String. Define facet for plotting.}

\item{faceted_by_2}{String. Define facet for plotting.}

\item{jitter_width}{Double. Define how big is the with of the scatter points.}

\item{.alpha}{Double. Define transparency for your scatter points.}

\item{.dot_size}{Double. Define the size of the scatter points.}

\item{...}{Other parameters to pass to the function.}
}
\value{

}
\description{
This function makes a boxplot and jitter plot for animal comparisons.
}
\examples{
# # The example is still missing...
}
