% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_cum_occu_wave_plot_func.R
\name{my_cum_occu_wave_plot_func}
\alias{my_cum_occu_wave_plot_func}
\title{my_cum_occu_wave_plot_func}
\usage{
my_cum_occu_wave_plot_func(my_dataset, reffer_wave_thres)
}
\arguments{
\item{my_dataset}{A dataframe with raw data of waves.}

\item{reffer_wave_thres}{A double. Reference value i n seconds indicating the threshold for the calculus of cumulative waves occurrence.}
}
\value{
A ggplot object (list) ploting cumulative occurrence of waves in the different animals groups.
}
\description{
plot cum occurence of waves
}
\examples{
# The example is still missing...
}
