% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summ_stat_func.R
\name{my_summ_stat_func}
\alias{my_summ_stat_func}
\title{Summary statistics of multiples variables}
\usage{
my_summ_stat_func(
  my_dataset,
  my_grouping_vars = c("Animal", "Treatment", "Condition"),
  Na_rm = TRUE
)
}
\arguments{
\item{my_dataset}{A previously cleaned dataframe with the function \code{clean_and_tidy_df_func()}.}

\item{my_grouping_vars}{Character vector. A character vector of groups names assigned to perform the aggregation. Don't change at least you know what you are doing!}

\item{Na_rm}{Boolean. Shall Nas be removed for summary functions computation?. Default to TRUE.}
}
\value{
Return a dataframe with the mean, median, SD, SEM, number of elements analized (cells or wave), number of animals, and asses Normality of the distribution with the Shapiro-Wilk test.
}
\description{
Compute the summary statistics of the multiples variables from the wave analysis of a previously cleaned dataset. Compute the \code{mean}, \code{sd}, \code{sem}, \code{median}, \code{n_Waves}, \code{n_Cells}, \code{n_Animals} and \code{Normality_Shapiro_p.} NOTE: Shapiro-Wilk test will fail if less than 2 samples!.
}
\examples{
# the example is missing
}
