% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnlysisOfWaves-package.R, R/utils-pipe.R
\docType{package}
\name{\%>\%}
\alias{AnlysisOfWaves}
\alias{AnlysisOfWaves-package}
\alias{\%>\%}
\title{AnlysisOfWaves: Clean, tidy, asses statistics and plot data from Ca2+ waves in the SERCA publication}
\usage{
lhs \%>\% rhs
}
\description{
What the package does (one paragraph).

See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
\author{
\strong{Maintainer}: Ruben Lopez \email{rjlopez2@gmail.com} (\href{https://orcid.org/0000-0002-3802-3158}{ORCID})

}
\keyword{internal}
