% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_func.R
\name{sem_func}
\alias{sem_func}
\title{Standard Error of the Mean (SEM)}
\usage{
sem_func(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{should the NA values be removed?}
}
\value{
The SEM from a given vector
}
\description{
Standard Error of the Mean (SEM)
}
\examples{

x <- c(0:10)
sem_x <- sem(x)

}
