% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superplot_func.R
\name{superplot_func}
\alias{superplot_func}
\title{Superplot with violin and scatter plot}
\usage{
superplot_func(
  dataset,
  yaxe,
  base_violin = "cells",
  dot_layer = "cells",
  animal_layer = TRUE,
  wave_size = 2,
  wave_alpha = 0.4,
  cell_size = 2.5,
  cell_alpha = 0.4,
  animal_size = 5,
  animal_alpha = 0.6,
  base_font_size = 15,
  jitter_width_wave = 3,
  jitter_width_cell = 3,
  jitter_width_animal = 3,
  ...
)
}
\arguments{
\item{dataset}{A dataframe object. The dataset for analysis, eg wave kinetics or wave occurrence df.}

\item{yaxe}{A string. The variable name of interest to plot.}

\item{base_violin}{A string. Which base plot would you like to plot as a violin plot?. Options are: 1. \code{waves}, 2. \code{cells}. Default to \code{cells}.}

\item{dot_layer}{A string. Which data would you like to see in the next layer?. Options are: 1. \code{waves}, 2. \code{cells}, 3. \code{waves_and_cells}, 4. \code{none}. Default to \code{cells}.}

\item{animal_layer}{A logic value. Would you like to add the Animal dot layer? default to \code{TRUE}.}

\item{wave_size}{A double. Value assigned to the size of the dots. Default to \code{2}.}

\item{wave_alpha}{A double. Value assigned to the transparency of the dots. Default to \code{0.4}.}

\item{cell_size}{A double. Value assigned to the size of the dots. Default to \code{2.5}.}

\item{cell_alpha}{A double. Value assigned to the transparency of the dots. Default to \code{0.4}.}

\item{animal_size}{A double. Value assigned to the size of the dots. Default to \code{5}.}

\item{animal_alpha}{A double. Value assigned to the transparency of the dots. Default to \code{0.6}.}

\item{base_font_size}{A integer. Modify the size of the fonts. Default to 15.}

\item{jitter_width_wave}{Double. Determine the size of spread of the scatter points for waves.}

\item{jitter_width_cell}{Double. Determine the size of spread of the scatter points for cells.}

\item{jitter_width_animal}{Double. Determine the size of spread of the scatter points for animal.}

\item{...}{Pass additional parameters to the internla plot layers.}
}
\value{
A ggplot object with a superplot indicating multiples layers of  Animals, cell or waves.
}
\description{
Make a violin and scatter plot with aggregated data at different levels (animal, cell or waves). The layers are constructed in this order: (wave/cell) violin layer -> (wave/cell) small dot layer -> (Animal) large dot layer. Note: this function is depending on other functions to build the multiples layers up.
}
\examples{
# the example is missing
}
