% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wave_layer_vio_ggplot_func.R
\name{wave_layer_vio_ggplot_func}
\alias{wave_layer_vio_ggplot_func}
\title{wave_layer_vio_ggplot_func}
\usage{
wave_layer_vio_ggplot_func(
  dataset,
  yaxe,
  xaxe = "Animal",
  faceted_by_1 = "Condition",
  faceted_by_2 = "."
)
}
\arguments{
\item{dataset}{A dataframe object. The dataset for analysis, eg wave kinetics or wave occurrence df.}

\item{yaxe}{A string. The variable name of interest to plot.}

\item{xaxe}{A string. The variable you want to compare with. Default to \code{Animal}.}

\item{faceted_by_1}{A string. Grouping variables for visualization. Default to \code{Condition}.}

\item{faceted_by_2}{A string. Aditional grouping variables for visualization. Default to \code{.}.}
}
\value{
A ggplot object with a violin plot canvas to add additional layers.
}
\description{
Create a base canvas violin ggplot  object. This function make single waves violin plot with the raw data that is, no aggregation on single waves for multiples cells. Note: this function is wrapped inside the final superplot function.
}
\examples{
# the example is missing
}
