% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcxntest_by_group_multivar_func.R
\name{wilcxntest_by_group_multivar_func}
\alias{wilcxntest_by_group_multivar_func}
\title{wilcxntest_by_group_multivar_func}
\usage{
wilcxntest_by_group_multivar_func(
  my_dataset,
  my_var_set,
  group_1 = "Condition",
  group_2 = "Animal"
)
}
\arguments{
\item{my_dataset}{A cleaned dataframe.}

\item{my_var_set}{A character vector with the names of variables to test.}

\item{group_1}{Character. the name for the first grouping comparison variable.}

\item{group_2}{Character. the name for the second grouping comparison variable.}
}
\value{
a datatable with the output of the Wilcoxon test
}
\description{
Wilcoxon test for multiples comparison
}
\examples{
# no example made jet.
}
